requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.21
click==8.1.3
clickhouse-connect==0.5.22
cmake==3.26.3
duckdb==0.7.1
exceptiongroup==1.1.1
fastapi==0.95.1
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.4.0
gTTS==2.3.1
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
huggingface-hub==0.14.1
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.2.0
lit==16.0.2
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.1
packaging==21.3
pandas==2.0.1
Pillow==9.5.0
pip-autoremove==0.10.0
pipdeptree==2.7.0
pluggy==1.0.0
posthog==3.0.1
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyglet==2.0.5
Pygments==2.14.0
pyparsing==3.0.9
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==13.3.2
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.98
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.26.1
sympy==1.11.1
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.14
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
wikipedia==1.4.0
yarl==1.8.2
zstandard==0.21.0
